+++ apps/backend/src/libs/hooks/hooks.ts
@@ -1,1 +1,1 @@
-export { checkUserPermissions } from "./check-user-permissions.hook.js";
+export { checkUserPermissions } from "./check-user-permissions/check-user-permissions.hook.js";