+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -61,7 +61,11 @@ import {
 } from "~/modules/vendors/vendors.js";
 
 import { storage } from "../storage/storage.js";
-import { chatSocket, handleError } from "./middlewares/middlewares.js";
+import {
+	chatSocket,
+	handleError,
+	notificationsSocket,
+} from "./middlewares/middlewares.js";
 
 type RootReducer = {
 	activities: ReturnType<typeof activitiesReducer>;
@@ -120,7 +124,11 @@ class Store {
 					thunk: {
 						extraArgument: this.extraArguments,
 					},
-				}).prepend([handleError, chatSocket({ extra: this.extraArguments })]);
+				}).prepend([
+					handleError,
+					chatSocket({ extra: this.extraArguments }),
+					notificationsSocket({ extra: this.extraArguments }),
+				]);
 			},
 			reducer: {
 				activities: activitiesReducer,