+++ apps/frontend/src/libs/types/types.ts
@@ -3,8 +3,10 @@ export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
 export { type IconName } from "./icon-name.type.js";
 export { type MenuItem } from "./menu-item.type.js";
 export { type RootState } from "./root-state.type.js";
+export { type SelectOption } from "./select-option.type.js";
 export { type ValueOf } from "./value-of.type.js";
 export {
+	type PaginationRequestDto,
 	type PaginationResponseDto,
 	type ServerErrorDetail,
 	type ServerErrorResponse,