+++ apps/frontend/src/modules/user-notifications/slices/actions.ts
@@ -1,4 +1,4 @@
-import { createAsyncThunk } from "@reduxjs/toolkit";
+import { createAction, createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
 
@@ -10,12 +10,12 @@ import { name as sliceName } from "./user-notifications.slice.js";
 
 const getUserNotifications = createAsyncThunk<
 	AllNotificationsResponseDto,
-	undefined,
+	string | undefined,
 	AsyncThunkConfig
->(`${sliceName}/get-user-notifications`, (_, { extra }) => {
+>(`${sliceName}/get-user-notifications`, (search = "", { extra }) => {
 	const { userNotificationsApi } = extra;
 
-	return userNotificationsApi.getUserNotifications();
+	return userNotificationsApi.getUserNotifications(search);
 });
 
 const checkHasUserUnreadNotifications = createAsyncThunk<
@@ -46,8 +46,22 @@ const setReadNotifications = createAsyncThunk<
 	},
 );
 
+const joinRoom = createAction(`${sliceName}/join-room`, (userId: string) => {
+	return {
+		payload: userId,
+	};
+});
+
+const leaveRoom = createAction(`${sliceName}/leave-room`, (userId: string) => {
+	return {
+		payload: userId,
+	};
+});
+
 export {
 	checkHasUserUnreadNotifications,
 	getUserNotifications,
+	joinRoom,
+	leaveRoom,
 	setReadNotifications,
 };