+++ apps/frontend/src/modules/users/libs/enums/enums.ts
@@ -1,1 +1,1 @@
-export { UsersApiPath } from "@trackmates/shared";
+export { UserSex, UsersApiPath } from "@trackmates/shared";