+++ packages/shared/src/index.ts
@@ -1,6 +1,8 @@
 export {
 	EMPTY_ARRAY_LENGTH,
+	LAST_ARRAY_ITEM,
 	MAX_FILE_SIZE_IN_MB,
+	PREVIOUS_INDEX_OFFSET,
 } from "./libs/constants/constants.js";
 export {
 	APIPath,
@@ -15,6 +17,10 @@ export {
 } from "./libs/enums/enums.js";
 export { ValidationError } from "./libs/exceptions/exceptions.js";
 export {
+	checkIsDatePrecedesAnotherByOneDay,
+	checkIsThisYear,
+	checkIsToday,
+	checkIsYesterday,
 	configureString,
 	getDifferenceInHours,
 	getFormattedDate,
@@ -59,6 +65,10 @@ export {
 	activityCreateFinishSectionValidationSchema,
 	activityDeleteFinishSectionValidationSchema,
 } from "./modules/activities/activities.js";
+export {
+	type ActivityLikeRequestDto,
+	activityLikeChangeValidationSchema,
+} from "./modules/activity-likes/activity-likes.js";
 export { AuthApiPath, AuthError } from "./modules/auth/auth.js";
 export {
 	type ChatMessageCreateRequestDto,
@@ -133,6 +143,7 @@ export {
 	type FriendUnfollowRequestDto,
 	FriendsApiPath,
 	addFriendValidationSchema,
+	friendGetAllQueryValidationSchema,
 	friendIdParameterValidationSchema,
 } from "./modules/friends/friends.js";
 export {
@@ -194,6 +205,7 @@ export {
 	type UserGetAllResponseDto,
 	type UserGetByIdRequestDto,
 	type UserProfileRequestDto,
+	UserSex,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
@@ -201,6 +213,7 @@ export {
 	UsersApiPath,
 	userIdParametersValidationSchema,
 	userProfileValidationSchema,
+	userSexToPronoun,
 	userSignInValidationSchema,
 	userSignUpValidationSchema,
 } from "./modules/users/users.js";