+++ packages/shared/src/libs/helpers/check-is-this-year/check-is-this-year.helper.ts
@@ -1,0 +1,7 @@
+import { isThisYear } from "date-fns";
+
+const checkIsThisYear = (date: Date | number | string): boolean => {
+	return isThisYear(date);
+};
+
+export { checkIsThisYear };