+++ packages/shared/src/modules/users/libs/types/user-auth-response-dto.type.ts
@@ -1,4 +1,6 @@
+import { type ValueOf } from "../../../../libs/types/types.js";
 import { type GroupResponseDto } from "../../../groups/groups.js";
+import { type UserSex } from "../enums/enums.js";
 
 type UserAuthResponseDto = {
 	avatarUrl: null | string;
@@ -9,6 +11,7 @@ type UserAuthResponseDto = {
 	id: number;
 	lastName: string;
 	nickname: null | string;
+	sex: ValueOf<typeof UserSex> | null;
 	updatedAt: string;
 };